fix(export): fix the nil go context passed to `archives.FilesFromDisk` for exporting package archive
When using the new version package `github.com/mholt/archives` to generate archive,
it requires a parameter as context. We pass nil as context, which can cause panic when
exec `pkg export`. We now fix it by passing `context.Background()` as the context.

related commit for this bug: a0d398e
related PR for this bug: #22

Signed-off-by: genshen <genshenchu@gmail.com>