feat(fetch): download package to a temp dir and then move it to gloabl cache dir
In pkg's cache strategy, if the global cache dir exists, the pkg downloading step will skip
this package and use the global cache dir.
Before this commit, pkg downloads package source to directly to the global cache dir. If the
user interrupts the downloading (e.g. CTR+C), the global cache dir for the package source code
can be empty and errors may occurs in the building step.
In this commit, we download the package to a temp dir first and move to the global
cache dir. The new downloading strategy can avoid the empty package source code dir.